make inference.py dataset-agnostic; add use_difficult option to voc dataset
