+++ maskrcnn_benchmark/data/datasets/evaluation/coco/__init__.py
@@ -1,0 +1,17 @@
+from .coco_eval import do_coco_evaluation
+
+
+def coco_evaluation(dataset,
+                    predictions,
+                    output_folder,
+                    box_only,
+                    iou_types,
+                    expected_results,
+                    expected_results_sigma_tol):
+    return do_coco_evaluation(dataset=dataset,
+                              predictions=predictions,
+                              box_only=box_only,
+                              output_folder=output_folder,
+                              iou_types=iou_types,
+                              expected_results=expected_results,
+                              expected_results_sigma_tol=expected_results_sigma_tol)
