+++ maskrcnn_benchmark/data/datasets/evaluation/voc/__init__.py
@@ -1,0 +1,18 @@
+import logging
+
+from .voc_eval import eval_detection_voc, do_voc_evaluation
+
+
+def voc_evaluation(dataset,
+                   predictions,
+                   output_folder,
+                   box_only,
+                   **_):
+    logger = logging.getLogger("maskrcnn_benchmark.inference")
+    if box_only:
+        logger.warning("voc evaluation doesn't support box_only, ignored.")
+    logger.info("performing voc evaluation, ignored iou_types.")
+    return do_voc_evaluation(dataset=dataset,
+                             predictions=predictions,
+                             output_folder=output_folder,
+                             logger=logger)
