+++ tools/train_net.py
@@ -84,17 +84,18 @@ def test(cfg, model, distributed):
     if cfg.MODEL.MASK_ON:
         iou_types = iou_types + ("segm",)
     output_folders = [None] * len(cfg.DATASETS.TEST)
+    dataset_names = cfg.DATASETS.TEST
     if cfg.OUTPUT_DIR:
-        dataset_names = cfg.DATASETS.TEST
         for idx, dataset_name in enumerate(dataset_names):
             output_folder = os.path.join(cfg.OUTPUT_DIR, "inference", dataset_name)
             mkdir(output_folder)
             output_folders[idx] = output_folder
     data_loaders_val = make_data_loader(cfg, is_train=False, is_distributed=distributed)
-    for output_folder, data_loader_val in zip(output_folders, data_loaders_val):
+    for output_folder, dataset_name, data_loader_val in zip(output_folders, dataset_names, data_loaders_val):
         inference(
             model,
             data_loader_val,
+            dataset_name=dataset_name,
             iou_types=iou_types,
             box_only=cfg.MODEL.RPN_ONLY,
             device=cfg.MODEL.DEVICE,
